requirements.txt
pandas>=2.0.0
requests>=2.31.0
pydantic>=2.5.0
html2text>=2020.1.16
openai>=1.3.0
python-dotenv>=1.0.0 